requirements.txt
pip==24.1
cryptography==42.0.8
mysql-connector-python==8.4.0
pydrive2==1.19.0

altgraph==0.17.4
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
Eel==0.16.0
future==1.0.0
gevent==24.2.1
gevent-websocket==0.10.1
google-api-core==2.19.0
google-api-python-client==2.134.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.1
greenlet==3.0.3
httplib2==0.22.0
idna==3.7
mysql-connector-python==8.4.0
oauth2client==4.1.3
packaging==24.1
pefile==2023.2.7
proto-plus==1.24.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyDrive2==1.19.0
pyinstaller==6.8.0
pyinstaller-hooks-contrib==2024.7
pyOpenSSL==24.1.0
pyparsing==3.1.2
pywin32-ctypes==0.2.2
PyYAML==6.0.1
requests==2.32.3
rsa==4.9
setuptools==70.1.0
six==1.16.0
uritemplate==4.1.1
urllib3==2.2.2
whichcraft==0.6.1
zope.event==5.0
zope.interface==6.4.post2
